+++ client/flutter/lib/components/carousel/carousel_slide.dart
@@ -1,83 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:page_view_indicator/page_view_indicator.dart';
 import 'package:html/dom.dart' as dom;
-
-class CarouselView extends StatelessWidget {
-  final List<CarouselSlide> items;
-
-  CarouselView({@required this.items});
-
-  final pageIndexNotifier = ValueNotifier<int>(0);
-
-  final PageController pageController = PageController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: <Widget>[
-        ShaderMask(
-          shaderCallback: (Rect bounds) {
-            return LinearGradient(
-                begin: Alignment(0.0, 0.0),
-                end: Alignment(0.0, 1.0),
-                colors: <Color>[Colors.white, Colors.white, Colors.transparent],
-                stops: [0.0, 0.5, 0.6]).createShader(bounds);
-          },
-          child: PageView(
-            controller: pageController,
-            onPageChanged: (i) => pageIndexNotifier.value = i,
-            children: this.items,
-          ),
-        ),
-        Align(
-          alignment: FractionalOffset.bottomCenter,
-          child: SafeArea(
-            child: Container(
-                padding: EdgeInsets.only(bottom: 8),
-                child: _buildPageViewIndicator(context)),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildPageViewIndicator(BuildContext context) {
-    var width = MediaQuery.of(context).size.width;
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Container(
-          constraints: BoxConstraints(maxWidth: width * 0.5),
-          child: FittedBox(
-            child: PageViewIndicator(
-              pageIndexNotifier: pageIndexNotifier,
-              length: this.items.length,
-              normalBuilder: (animationController, index) => Circle(
-                size: 20.0,
-                color: Color(0x99FFFFFF),
-              ),
-              highlightedBuilder: (animationController, index) =>
-                  ScaleTransition(
-                scale: CurvedAnimation(
-                  parent: animationController,
-                  curve: Curves.ease,
-                ),
-                child: Circle(
-                  size: 28.0,
-                  color: Colors.white,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
 class CarouselSlide extends StatefulWidget {
   final SvgPicture graphic;
   final String title;
