+++ client/flutter/lib/pages/facts_carousel_page.dart
@@ -1,5 +1,7 @@
 import 'package:WHOFlutter/api/content/dynamic_content.dart';
-import 'package:WHOFlutter/components/carousel.dart';
+import 'package:WHOFlutter/components/carousel/carousel.dart';
+import 'package:WHOFlutter/components/carousel/carousel_slide.dart';
+
 import 'package:WHOFlutter/components/page_scaffold/page_header.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
